% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Diag_Scaling_Mat_Sparse.R
\name{adj_to_probTrans}
\alias{adj_to_probTrans}
\title{Adjacency to Probability Transition Matrix}
\usage{
adj_to_probTrans(mat)
}
\arguments{
\item{mat}{A matrix like object (either a matrix, sparse matrix or dataframe)}
}
\value{
the function returns a matrix of the form dgCMatrix from
from the Matrix package, wrap in as.matrix() if necessary
}
\description{
Takes an Adjacency matrix and scales each column to 1 or 0.
}
\details{
The returned matrix will be such that each entry A\link{i,j} describes the
probability of travelling from vertex j to vertex i during a random
walk. (Note that column -> row is the transpose of what igraph returns)
which is row to column)
}
\examples{
adj_to_probTrans(matrix(1:3, 3))


}
