% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne_network.R
\name{aracne_network}
\alias{aracne_network}
\title{Re-format ARACNe Network.}
\usage{
aracne_network(regulon)
}
\arguments{
\item{regulon}{(regulon) ARACNe network.}
}
\value{
(data.frame) Network, with "reg", "target" and "direction" in column name. For direction, 1/0 denotes positive/negative regulation.
}
\description{
Re-format ARACNe network in regulon object to data.frame with regulator, target and direction columns.
}
\examples{
library(bcellViper)
data(bcellViper)
aracne_network(regulon[1:10])

}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
