% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Make Bubbleplot}
\usage{
bubble_plot(
  s_mat,
  c_mat,
  n_mat,
  col = colorRampPalette(c("Blue", "Grey", "Red"))(100),
  breaks = seq(-2, 2, length.out = 100),
  main = NULL
)
}
\arguments{
\item{s_mat}{(matrix) Matrix denotes the size of bubbles.}

\item{c_mat}{(matrix) Matrix denotes the color of bubbles.}

\item{n_mat}{(matrix) Matrix denotes the name of bubbles.}

\item{col}{(character) Colors.}

\item{breaks}{(numeric) Breakpoints of colors.}

\item{main}{(character) Title.}
}
\value{
(NULL) A bubbloplot.
}
\description{
Make bubbleplot.
}
\examples{
s_mat <- c_mat <- n_mat <- matrix(1:12, 3, 4, dimnames = list(1:3, 1:4))
bubble_plot(s_mat, c_mat, n_mat, breaks = seq(1, 12, length.out = 100), main = "Test")

}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
