% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Diag_Scaling_Mat_Sparse.R
\name{create_sparse_diag_sc_inv_mat}
\alias{create_sparse_diag_sc_inv_mat}
\title{Diagonal Factors of Sparse Matrix Inv}
\usage{
create_sparse_diag_sc_inv_mat(mat)
}
\arguments{
\item{mat}{Input Matrix, may either be a matrix, sparse matrix
dgCMatrix or probably even a data frame, as long as \code{nrow} and \code{colSum}
work on the data as expected it should work, output is always a
sparse matrix of class dgCMatrix

This should take the an adjacency matrix as input and the output
can be multiplied by the original matrix to scale it to 1.

A directed adjacency matrix should be such that A\link{i,j}
describes the the weight
of an edge from vertex j to vertex i (this is, unfourtunately
the transpose of what igraph gives back.)}
}
\value{
A diagonal matrix such that multiplication by the original
matrix will lead to all columns adding
to 1 (or 0 in the case of a column of zeros)
The output class is sparse matrix "dgCMatrix" from the Matrix package
use \code{as.matrix()} in order to get a matrix type, I didn't put that
Inside the function because it would have been confusing.
}
\description{
Takes a sparse matrix
and returns a diagonal matrix such that each column contains
either colsum() or 0. (D in the Papers)
multiplication with this matrix would have all columns
sum to 1 (or zero if that column already summed to zero)
}
\examples{
(mat <- matrix(1:9, nrow = 3))
mat \%*\% create_sparse_diag_scaling_mat(mat) \%>\% colSum()
mat \%*\% create_sparse_diag_scaling_mat(mat)

> [1] 1 1 1 1 1
}
