% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_bin.R
\name{gene_bin}
\alias{gene_bin}
\title{Bin Gene Expression Space}
\usage{
gene_bin(genes, expmat, sep = 5)
}
\arguments{
\item{genes}{(character) Genes to be analyzed.}

\item{expmat}{(matrix) Gene expression matrix.}

\item{sep}{(numeric) Number of bins.}
}
\value{
(matrix) Border values of gene expression bins.
}
\description{
Bin gene expression space for marginal/joint probability calculation.
}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
