% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color_gradient.R
\name{get_color_gradient}
\alias{get_color_gradient}
\title{Generate Color Gradient}
\usage{
get_color_gradient(
  x,
  col = colorRampPalette(c("Blue", "Grey", "Red"))(100),
  breaks = seq(-2, 2, length.out = 100)
)
}
\arguments{
\item{x}{(numeric) Vector based on which color gradient is generated.}
}
\value{
(character) Colors.
}
\description{
Generate color gradient for, e.g. gene expression.
}
\examples{
get_color_gradient(-2:2)

}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
