% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_dist.R
\name{P_dist}
\alias{P_dist}
\title{Calculate Probability-Based Distance}
\usage{
P_dist(pxy, pxpy, method = c("difference", "mi"))
}
\arguments{
\item{pxy}{(list) Joint probability generated by PXY fuction.}

\item{pxpy}{(list) Product of marginal probability generated by PXPY fuction.}

\item{method}{(character) Method for calculating distance, either PXY-PXPY
("difference") or mutual information ("mi").}
}
\value{
(list) Distance values.
}
\description{
Calculate distance based on PXY and PXPY.
}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
\keyword{internal}
