% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pagerank.R
\name{plot_graph}
\alias{plot_graph}
\title{Plots the dependency graph}
\usage{
plot_graph(mirror, top = 25, cex = 0.5, random.seed = 1)
}
\arguments{
\item{mirror}{CRAN mirror, e.g. http://cran.revolutionanalytics.com or https://mran.revolutionanalytics.com/snapshot/2016-03-15}

\item{top}{integer that indicates how many packages to include in the graph}

\item{cex}{character expansion ratio for plot}

\item{random.seed}{A random seed value. Used by igraph to initialise the plot layout. The function resets the seed to the original value, i.e. it will not have any impact on the random value stream in your R session.}
}
\description{
Plots the dependency graph
}

