% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Diag_Scaling_Mat_Sparse.R
\name{power_walk_prob_trans}
\alias{power_walk_prob_trans}
\title{Power Walk Probability Transition Matrix}
\usage{
power_walk_prob_trans(A, beta = 10)
}
\arguments{
\item{beta}{is the ratio of probability between following an edge or
making a random jump in a graph. If beta is 0 the probability of
following an edge is no more likely than a randum jump.
As beta is made larger, the probability of making random jumps becomes smaller.
TODO if beta is zero the trans mat should be uniform.
Value defaults to 10}
}
\value{
The output matrix will be of the class dgCmatrix
}
\description{
Takes an adjacency matrix and returns the transition
probability matrix using the Power Walk Method
}
