% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Diag_Scaling_Mat_Sparse.R
\name{power_walk_stationary_point}
\alias{power_walk_stationary_point}
\title{Power Walk Stationary Point}
\usage{
power_walk_stationary_point(A, beta = 10, eta)
}
\arguments{
\item{beta}{See \code{\link{power_walk_prob_trans}}

probability of travelling from vertex j to vertex i during a random
walk. (Note that column -> row is the transpose of what igraph returns)
which is row to column)}

\item{mat}{A matrix like object (either a matrix, sparse matrix or dataframe)}
}
\value{
the function returns a matrix of the form dgCMatrix from
from the Matrix package, wrap in as.matrix() if necessary
}
\description{
Takes an adjacency matrix and returns the stationary
and returns the stationary distribution using the power walk method
}
\details{
The returned vector will describe the probability of landing on any
given vertex of a graph during a random walk with a smoothing parameter.
}
\examples{
adj_to_probTrans(matrix(1:3, 3))


}
