% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PX.R
\name{PX}
\alias{PX}
\title{Marginal Probability Calculation}
\usage{
PX(expmat, bin)
}
\arguments{
\item{expmat}{(matrix) Gene expression matrix.}

\item{bin}{(matrix) Results of gene_bin function.}
}
\value{
(matrix) Marginal probability matrix.
}
\description{
Calculate marginal probability.
}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
\keyword{internal}
