% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PXPY.R
\name{PXPY}
\alias{PXPY}
\title{Product of Marginal Probability}
\usage{
PXPY(px, combinations)
}
\arguments{
\item{px}{(matrix) Marginal probability generated by PX fuction.}

\item{combinations}{(character) Variable combinations ("x_y") to be analyzed.}
}
\value{
(list) Product of marginal probability.
}
\description{
Calculate the product of marginal probability.
}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
