% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PXY.R
\name{PXY}
\alias{PXY}
\title{Joint Probability Calculation}
\usage{
PXY(expmat, bin, x, y)
}
\arguments{
\item{expmat}{(matrix) Gene expression matrix.}

\item{bin}{(matrix) Results of gene_bin function.}

\item{x}{(character) The first variable.}

\item{y}{(character) The pairing second variable.}
}
\value{
(list) Joint probability.
}
\description{
Calculate joint probability.
}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
