% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Diag_Scaling_Mat_Sparse.R
\name{random_surfer_stationary}
\alias{random_surfer_stationary}
\title{Random Surfer Stationary Distribution}
\usage{
random_surfer_stationary(A, alpha = 0.85, dp = 2)
}
\arguments{
\item{A}{An adjacancy matrix such that A\link{i,j} indicates travel j -> i}

\item{alpha}{the value of the smoothing parameter,
\itemize{
\item 0 would indicate that every vertex is teleported to during a random walk
\item 1 would indicate that teleporting can't happen
}}

\item{dp}{how many decimal points of accuracy to use TODO make eta}
}
\description{
Takes an Adjacency matrix and returns the Stationary Distribution using Random Surfer
}
\details{
The returned vector will describe the probability of landing at each
vertex during a random walk
Using the Random Surfer model.
}
\examples{
adj_to_probTrans(matrix(1:3, 3))


}
