% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility_network.R
\name{accessibility_network}
\alias{accessibility_network}
\title{Build Network from Accessibility Peaks.}
\usage{
accessibility_network(
  table,
  regulators,
  species = c("hg19", "mm9", "hg38", "mm10"),
  upstream = 2000,
  downstream = 200,
  p.cutoff = 5e-05,
  w = 7
)
}
\arguments{
\item{table}{(data.frame) Peaks, with "Chr", "Start" and "End" in column name, and peak ID in row names.}

\item{regulators}{(character) Regulators to be analyzed.}

\item{species}{(character) Could be "hg19", "hg38", "mm9" or "mm10".}

\item{upstream}{@param downstream (numeric) Promoter region from TSS for target identificaton}

\item{p.cutoff}{(numeric) P-value cutoff for motifs searching within peaks for TF identificaton.}

\item{w}{(numeric) Window size for motifs searching within peaks for TF identificaton.}
}
\value{
(data.frame) Network, with "reg" and "target" in column name.
}
\description{
Build network from accessibility, e.g. ATAC-Seq peaks.
}
\examples{
table <- data.frame(Chr=c("chr1", "chr1"), Start=c(713689, 856337), End=c(714685, 862152),
                    row.names=c("A", "B"), stringsAsFactors=F)
regulators=c("FOXF2", "MZF1")
accessibility_network(table, regulators, species = "hg19")

}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
