% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformation_network.R
\name{conformation_network}
\alias{conformation_network}
\title{Build Network from Conformation Peaks.}
\usage{
conformation_network(
  table,
  regulators,
  species = c("hg19", "mm9", "hg38", "mm10"),
  range = 500,
  upstream = 2000,
  downstream = 200,
  p.cutoff = 5e-05,
  w = 7
)
}
\arguments{
\item{table}{(data.frame) Records, with "Chr1", "Position1", "Strand1", "Chr2", "Position2" and "Strand2" in column name, and record ID in row names.}

\item{regulators}{(character) Regulators to be analyzed.}

\item{species}{(character) Could be "hg19", "hg38", "mm9" or "mm10".}

\item{range}{(numeric) Search region defined from "Position1" and "Position2".}

\item{upstream}{@param downstream (numeric) Promoter region defined by distance from TSS.}

\item{p.cutoff}{(numeric) P-value cutoff for motifs searching within peaks for TF identificaton.}

\item{w}{(numeric) Window size for motifs searching within peaks for TF identificaton.}
}
\value{
(data.frame) Network, with "reg" and "target" in column name.
}
\description{
Build network from conformation, e.g. HiChIP records.
}
\examples{
table <- data.frame(Chr1=c("chr1", "chr1"), Position1=c(569265, 713603), Strand1=c("+", "+"),
                    Chr2=c("chr4", "chr1"), Position2=c(206628, 715110), Strand2=c("+", "-"),
                    row.names=c("A", "B"), stringsAsFactors=F)
regulators=c("FOXF2", "MZF1")
conformation_network(table, regulators, species = "hg19")

}
\author{
DING, HONGXU (hd2326@columbia.edu)
}
